Rename auto-upgrade scripts to make uniform
