+++ src/auto-upgrade/updateConfig.js
