+++ src/auto-upgrade/updatePackage.js
